+++ apps/CMakeLists.txt
@@ -8,6 +8,8 @@ set (APPS_ROOT_DIR "${CMAKE_CURRENT_SOURCE_DIR}")
 
 set (APPS_SHARE_DIR "${CMAKE_CURRENT_BINARY_DIR}/share")
 
+collect (APP_INC_DIRS "${CMAKE_CURRENT_SOURCE_DIR}/include")
+
 add_subdirectory (machine)
 add_subdirectory (system)
 add_subdirectory (tests)
